+++ TowerDefense/Tower/Towertype/AttackerTower.swift
@@ -9,7 +9,7 @@ import Foundation
 class AttackerTower: Towers {
     var range: Int
     
-    init(name: String, hp: Int, price: Int, cd: Int, level: Int, position: (Int,Int), range:  Int) {
+    init(name: String, hp: Int, price: Int, cd: Int, level: Int, position: (Double,Double), range:  Int) {
         self.range = range
         super.init(name: name, hp: hp, price: price, cd: cd, level: level, position: position)
     }
